requirements.txt
robotframework==3.2.2
robotframework-pythonlibcore==2.1.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.5.0
selenium==3.141.0
urllib3==1.26.2

pip~=20.2.3
argparse~=1.4.0
Brotli~=1.0.9
cryptography~=3.1
lxml~=4.5.2
docutils~=0.16
wxPython~=4.0.7.post2
pillow~=8.0.0
PyYAML~=5.3.1
six~=1.15.0
setuptools~=49.2.1
Events~=0.2.2
requests~=2.24.0
jsonschema~=3.2.0